# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000/v1"
info:
  title: WASAText
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
tags:  # Add the tags array here
  - name: Messages
    description: Operations related to messages
  - name: Conversations
    description: Operations related to conversations
  - name: Get services
  - name: login
  - name: Profile
components:
  schemas:
    Conversation:
      title: conversation
      description: "This object represents a single conversation."
      type: object
      properties:
        id:
          description: Unique identifier.
          type: integer
          example: 1
          readOnly: true
          minimum: 1
          maximum: 999
        messages:
          type: array
          items: { $ref: '#/components/schemas/Message' }
          minItems: 0
          maxItems: 999
          description: "messages array"
        details:
          $ref: '#/components/schemas/ConversationDetails' # Use a named schema
      required:
        - id
        - messages
        - details
      
    ConversationDetails: # New named schema for 'details'
      oneOf:
        - $ref: '#/components/schemas/Chat'
        - $ref: '#/components/schemas/Group'
      description: "The type of conversation details"

    Group:
      properties:
        name:
          description: name of the group.
          type: string
          example: the koalas
          pattern: "^.+$"
          minLength: 3
          maxLength: 16
        description:
          description: "the description of the group."
          type: string
          pattern: "^.+$"
          minLength: 3
          maxLength: 256
        group_pic_url:
          type: string
          format: url
          description: "the URL of the group picture."
          minLength: 3
          maxLength: 64
        participants:
          description: the list of the group participants
          type: array
          items: { $ref: "#/components/schemas/User" }
          minItems: 2
          maxItems: 999
      required:
        - name
        - participants
    Chat:
      properties:
        chatters:
          type: array
          minItems: 2
          maxItems: 2
          description: "the list of the two chatters."
          items: { $ref: "#/components/schemas/User" }
      required:
        - chatters
    Message:
      title: message
      description: "This object represents a single message."
      type: object
      properties:
        id:
          description: Unique message identifier.
          type: integer
          example: 1
          readOnly: true
          minimum: 1
          maximum: 9999
        text:
          description: Text of this message
          pattern: "^.+$"
          minLength: 1
          maxLength: 256
          type: string
          example: "Hello, how are you?"
        photo_url:
          type: string
          format: url
          description: URL of the photo attached to the message.
          minLength: 3
          maxLength: 64
        received:
          description: list of the people that have received the message.
          type: array
          minItems: 0
          maxItems: 99
          items: 
            $ref: "#/components/schemas/User"
        read:
          description: list of the people that have read the message.
          type: array
          minItems: 0
          maxItems: 99
          items: 
            $ref: "#/components/schemas/User"
      required:
        - id
        - text
        - received
        - read
    ForwardMessage:
      type: object
      description: "This object represents a single forwarded message."
      properties:
        original_message_id:
          type: integer
          minimum: 1
          maximum: 9999
          description: The ID of the message being forwarded.
    Reaction:
      title: reaction
      description: "This object represents a single reaction."
      type: object
      properties:
        user_id:
          type: integer
          description: ID of the reacter.
          minimum: 1
          maximum: 999
        text:
          description: emoji Unicode.
          type: string
          pattern: "^.+$"
          minLength: 1
          maxLength: 1
          example: "🙂"
      required: [user_id, text]
    User:
      title: user
      description: "This object represents a single user."
      type: object
      properties:
        id:
          description: Unique user identifier.
          type: integer
          example: 1
          readOnly: true
          minimum: 1
          maximum: 999
        name:
          description: name of the user.
          type: string
          example: andre23Z
          pattern: "^.+$"
          minLength: 3
          maxLength: 16
        profile_pic_url:
          type: string
          format: url
          description: URL of the user's profile picture.
          maxLength: 64
    GoodRequest:
      description: "This object represents a successful request."
      type: object
      properties:
        message:
          description: "Success message"
          type: string
          example: "Request successful"
          pattern: "^.+$"
          minLength: 1
          maxLength: 64
    BadRequestError:
      type: object
      description : "This object represents a bad request error."
      properties:
        error:
          description: "Error type"
          type: string
          example: "Bad Request"
          pattern: "^.+$"
          minLength: 1
          maxLength: 32
        trace:
          type: string
          example: "Missing required field"
          description: "Error message"
          pattern: "^.+$"
          minLength: 1
          maxLength: 64
    InternalServerError:
      description: "This object represents an internal server error."
      type: object
      properties:
        error:
          description: "Error type"
          type: string
          example: "Internal Server Error"
          pattern: "^.+$"
          minLength: 1
          maxLength: 32
        trace:
          type: string
          example: "Database connection failed"
          description: "Error message"
          pattern: "^.+$"
          minLength: 1
          maxLength: 64
  responses:
    GoodRequest:
      description: The request was successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GoodRequest'
          example:
            message: "Request successful"
    BadRequest:
      description: The request was not compliant with the documentation (e.g., missing fields, etc.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
                
  securitySchemes:
    BearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes

  parameters:
    user_id:
      schema:
        type: integer
        example: 1
        readOnly: true
        minimum: 1
        maximum: 999
      name: user_id
      in: path
      required: true
      description: Unique identifier for a user.
      allowEmptyValue: false
    message_id:
      schema:
        type: integer
        example: 1
        readOnly: true
        minimum: 1
        maximum: 9999
      name: message_id
      in: path
      required: true
      description: Message identifier.
      allowEmptyValue: false
    conversation_id:
      schema:
        type: integer
        example: 27
        readOnly: true
        minimum: 1
        maximum: 999
      name: conversation_id
      in: path
      required: true
      description: Unique identifier for a conversation.
      allowEmptyValue: false
      
paths:
  /index:
    get:
      tags: ["Get services"]
      operationId: index
      summary: Get an index
      description: |-
        This is an example of an index API description
      responses:
        "200":
          {$ref: "#/components/responses/GoodRequest"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: User details
              properties:
                name:
                  type: string
                  description: The name of the user
                  example: 'maria'
                  pattern: '^[^\\n]*$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              example:
                identifier: "abcdef012345"
                message: "Login Successful"
              schema:
                description: User identifier
                type: object
                properties:
                  identifier:
                    # change here if you decide to use an integer
                    # or any other type of identifier
                    type: string
                    example: "abcdef012345"
                    minLength: 1
                    maxLength: 32
                    pattern: "^.+$"
                    description: User identifier
                  message:
                    type: string
                    example: "Login Successful"
                    pattern: "^.+$"
                    minLength: 1
                    maxLength: 32
                    description: Message
  /users/{user_id}/name:
    put:
      tags: ["Profile"]
      operationId: setMyUserName
      description: |-
        This endpoint allows the user to change their username.
        The new username must be unique.
      summary: Modify my Username
      parameters:
        - $ref: "#/components/parameters/user_id"
      requestBody:
        description: New username to be set
        content:
          application/json:
            schema:
              description: New username
              type: object
              properties:
                username:
                  pattern: "^[^\\n]*$"
                  description: The new username
                  type: string
                  example: "newUsername123"
                  minLength: 3
                  maxLength: 16
              required:
                - username
      responses:
        "200":
          description: Username updated successfully
          content:
            application/json:
              example:
                message: "Username updated successfully"
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: User not found
        "500": { $ref: "#/components/responses/InternalServerError" }

  /users/{user_id}/photo:
    put:
      tags: ["Profile"]
      operationId: setMyPhoto
      summary: Change my photo
      description: |-
        This endpoint allows the user to change their profile picture.
        The new photo must be a valid URL.
      parameters:
        - $ref: "#/components/parameters/user_id"
      requestBody:
        description: New photo to be set
        content:
          application/json:
            schema:
              description: New photo
              type: object
              properties:
                photo:
                  description: The new photo URL
                  minLength: 3
                  maxLength: 64
                  type: string
                  format: url
              required:
                - photo
      responses:
        "200":
          description: Profile photo updated successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /chat/{conversation_id}:
    parameters:
      - $ref: "#/components/parameters/conversation_id"
    post:
      operationId: addToGroup
      summary: Add one or more users to a group
      description: Adds users to the specified group.
      tags: ["Conversations"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: User IDs to be added to the group
              type: object
              properties:
                username:
                  description: user to be added to the group
                  type: string
                  example: "newUsername123"
                  pattern: "^[^\\n]*$"
              required:
                - username
      responses:
        "200":
          description: Users added successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Unauthorized (missing or invalid token)
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["Conversations"]
      operationId: leaveGroup
      summary: Leave a group
      description: Remove partecipation from a group.
      responses:
        "204":
          description: Message deleted successfully.
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /chat/{conversation_id}/name:
    put:
      description: Change the group name
      operationId: setGroupName
      summary: Set a group name
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      tags: ["Conversations"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: New group name
              type: object
              properties:
                name:
                  minLength: 3
                  maxLength: 16
                  pattern: "^.+$"
                  type: string
                  description: The new name of the group
              required:
                - name
      responses:
        "200":
          description: Group name modified successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Unauthorized (missing or invalid token)
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  
  /chat/{conversation_id}/photo:
    put:
      description: Change the group photo
      tags: ["Conversations"]
      operationId: setGroupPhoto
      summary: Change group photo
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      requestBody:
        description: New photo to be set
        content:
          application/json:
            schema:
              description: New photo
              type: object
              properties:
                photo:
                  maxLength: 64
                  type: string
                  format: url
              required:
                - photo
      responses:
        "200":
          description: Group photo updated successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /chat/{conversation_id}/messages:
    parameters:
      - $ref: "#/components/parameters/conversation_id"
    get:
      operationId: getConversation
      summary: Get messages from a conversation
      description: Retrieves all messages from the specified conversation.
      tags: ["Conversations"]
      responses:
        "200":
          description: List of private chat messages
          content:
            application/json:
              example:
                id: 1
                messages:
                  - id: 1
                    text: "Hello, how are you?"
                    photo_url: "https://example.com/photo.jpg"
                    received:
                      - id: 1
                        name: "andre23Z"
                        profile_pic_url: "https://example.com/profile.jpg"
                    read:
                      - id: 1
                        name: "andre23Z"
                        profile_pic_url: "https://example.com/profile.jpg"
                  - id: 2
                    text: "I'm fine, thank you!"
                    photo_url: "https://example.com/photo.jpg"
                    received:
                      - id: 2
                        name: "maria"
                        profile_pic_url: "https://example.com/profile.jpg"
                    read:
                      - id: 2
                        name: "maria"
                        profile_pic_url: "https://example.com/profile.jpg"
                details:
                  chatters:
                    - id: 1
                      name: "andre23Z"
                      profile_pic_url: "https://example.com/profile.jpg"
                    - id: 2
                      name: "maria"
                      profile_pic_url: "https://example.com/profile.jpg"
              schema:
                $ref: '#/components/schemas/Conversation'
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Conversation not found
        "401":
          description: Unauthorized (missing or invalid token)
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Send a message in a conversation
      description: Post a new message in a specific conversation (either private or group).
      operationId: sendMessage
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"  # Link to Message schema
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              example:
                id: 1
                text: "Hello, how are you?"
                message: "Message sent successfully"
                received: []
                read: []
              schema:
                $ref: "#/components/schemas/Message"  # Same Message schema to return the sent message
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Unauthorized (missing or invalid token)
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"

  /chat/{conversation_id}/forward:
    post:
      summary: Forward an existing message
      operationId: forwardMessage
      description: Forward an existing message to the specified conversation.
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardMessage"
      responses:
        "200":
          description: Message sent successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Unauthorized (missing or invalid token)
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"

  /chat/{conversation_id}/messages/{message_id}:
    delete:
      tags: ["Messages"]
      operationId: deleteMessage
      summary: Delete a message
      description: Deleting a message from a conversation.
      parameters:
        - $ref: "#/components/parameters/conversation_id"
        - $ref: "#/components/parameters/message_id"
      responses:
        "204":
          description: Message deleted successfully.
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /chat/{conversation_id}/messages/{message_id}/reaction:
    parameters:
      - $ref: "#/components/parameters/conversation_id"
      - $ref: "#/components/parameters/message_id"
    post:
      operationId: commentMessage
      summary: Reacting to a message
      description: Reacting a message of the specified conversation.
      tags: ["Messages"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reaction"
      responses:
        "200":
          description: Message sent successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Unauthorized (missing or invalid token)
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["Messages"]
      operationId: uncommentMessage
      summary: Unfollow another user
      description: Deleting the reaction to a message
      responses:
        "204":
          description: Reaction deleted successfully.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { description: User not found }
        "500": { $ref: "#/components/responses/InternalServerError" }
    
  /stream:
    get:
      tags:
        - "Get services"
      operationId: getMyConversations
      summary: Get personal conversations stream
      description: Retrieve a stream of the authenticated user's conversations.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Chats retrieved successfully
          content:
            application/json:
              example:
                conversations: []
              schema:
                description: List of conversations
                type: object
                properties:
                  conversations:
                    type: array
                    description: List of conversations
                    minItems: 0
                    maxItems: 999
                    items:
                      $ref: '#/components/schemas/Conversation'
                required: [conversations]
        "401":
          description: Unauthorized access
        "500":
          $ref: "#/components/responses/InternalServerError"